Add files via upload
A polling network may be considered as a multiple-queue with a cyclic server. Because of polling, the channel time for each station is comprised of a polling interval and a data transmission interval. For performance analysis, we obtain expressions for the mean waiting time of exhaustive, partially gated, and fully gated polling networks. From these mean waiting times, the corresponding average transfer delays are also obtained.